+++ tests/unit_tests/Commands/Remote/DummyCommand.php
@@ -11,22 +11,14 @@ use Pantheon\Terminus\Commands\Remote\SSHBaseCommand;
  */
 class DummyCommand extends SSHBaseCommand
 {
+    /**
+     * @var string
+     */
     protected $command = 'dummy';
-
-    protected $unavailable_commands = [
-        'avoided'        => 'alternative',
-        'no-alternative' => '',
-    ];
-
-    protected $valid_frameworks = [
-        'framework-a',
-        'framework-b',
-    ];
 
     public function dummyCommand($site_env_id, array $dummy_args)
     {
         $this->prepareEnvironment($site_env_id);
-
         return $this->executeCommand($dummy_args);
     }
 }